+++ packages/mobile/CHANGELOG.md
@@ -8,6 +8,18 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 8.21.5 (11/19/2025 PST)
+
+#### 🐞 Fixes
+
+- Auto-fill DatePicker width. [[#118](https://github.com/coinbase/cds/pull/118)] [DX-4988]
+
+## 8.21.4 (11/18/2025 PST)
+
+#### 🐞 Fixes
+
+- Update styling of DefaultSelectControl to use compact chips. [[#156](https://github.com/coinbase/cds/pull/156)]
+
 ## 8.21.3 (11/17/2025 PST)
 
 #### 🐞 Fixes